+++ packages/@expo/cli/src/export/exportApp.ts
@@ -42,6 +42,11 @@ export async function exportAppAsync(
     dumpSourcemap,
   }: Pick<Options, 'dumpAssetmap' | 'dumpSourcemap' | 'dev' | 'clear' | 'outputDir' | 'platforms'>
 ): Promise<void> {
+  // Set the environment to production or development
+  // lots of tools use this to determine if they should run in a dev mode.
+  process.env.NODE_ENV = dev ? 'development' : 'production';
+  process.env.BABEL_ENV = dev ? 'development' : 'production';
+
   const exp = await getPublicExpoManifestAsync(projectRoot);
 
   const publicPath = path.resolve(projectRoot, env.EXPO_PUBLIC_FOLDER);