+++ packages/@expo/cli/src/start/startAsync.ts
@@ -69,6 +69,11 @@ export async function startAsync(
 ) {
   Log.log(chalk.gray(`Starting project at ${projectRoot}`));
 
+  // Set the environment to production or development
+  // lots of tools use this to determine if they should run in a dev mode.
+  process.env.NODE_ENV = options.dev ? 'development' : 'production';
+  process.env.BABEL_ENV = options.dev ? 'development' : 'production';
+
   const { exp, pkg } = profile(getConfig)(projectRoot);
 
   const platformBundlers = getPlatformBundlers(exp);